requirements.txt
setuptools>6
django>=1.8,<1.9
djangorestframework>=2.4.0
oic>=0.7.6

-e git+http://github.com/jhuapl-boss/django-oidc.git#egg=django-oidc
drf-oidc-auth
